update: 更新了博客可持久化的功能

+++ server/src/main/java/com/example/NBUACM/entity/Blog.java
@@ -11,5 +11,5 @@ public class Blog {
     String uid;
     String username;
     String BlogContent;
-    int ProblemId;
+    Long ProblemId;
 }
